updating dim customer staging code

+++ sql_scripts/1. staging_customer_dim.sql
@@ -1,5 +1,6 @@
 --create if not exist
-DROP TABLE computer_std.dbo.dim_customer
+IF OBJECT_ID('computer_std.dbo.dim_customer', 'U') IS NOT NULL
+    DROP TABLE computer_std.dbo.dim_customer;
 
 CREATE TABLE computer_std.dbo.dim_customer (
     Customer_ID INT IDENTITY(1,1) PRIMARY KEY, 
@@ -7,12 +8,13 @@ CREATE TABLE computer_std.dbo.dim_customer (
     [Customer Surname] VARCHAR(250),
     [Customer Contact Number] NVARCHAR(250),
     [Customer Email Address] VARCHAR(250)
-)
+);
 --2. insert data into the table
-INSERT INTO computer_std.dbo.dim_customer ([Customer Name], [Customer Surname], [Customer Contact Number], [Customer Email Address])
-SELECT DISTINCT [Customer Name], [Customer Surname], MAX([Customer Contact Number]), MAX([Customer Email Address])
-FROM computer_std.dbo.pc_data GROUP BY [Customer Name], [Customer Surname];
-
+INSERT INTO computer_std.dbo.dim_customer 
+    ([Customer Name], [Customer Surname], [Customer Contact Number], [Customer Email Address])
+SELECT DISTINCT 
+    [Customer Name], [Customer Surname], [Customer Contact Number], [Customer Email Address]
+FROM computer_std.dbo.pc_data;
 
 --3. select all the data from dim table
 SELECT * FROM computer_std.dbo.dim_customer
